feat: migrate aws config to dtctl SDK
