+++ cmd/install.go
@@ -9,6 +9,7 @@ import (
 	"github.com/dynatrace-oss/dtwiz/pkg/analyzer"
 	"github.com/dynatrace-oss/dtwiz/pkg/featureflags"
 	"github.com/dynatrace-oss/dtwiz/pkg/installer"
+	awspkg "github.com/dynatrace-oss/dtwiz/pkg/installer/aws"
 	"github.com/dynatrace-oss/dtwiz/pkg/installer/azure"
 	"github.com/dynatrace-oss/dtwiz/pkg/installer/gcp"
 	k8s "github.com/dynatrace-oss/dtwiz/pkg/installer/kubernetes"
@@ -280,19 +281,14 @@ var installAWSCmd = &cobra.Command{
 	Short: "Set up Dynatrace AWS CloudFormation integration",
 	Args:  cobra.NoArgs,
 	RunE: func(cmd *cobra.Command, args []string) error {
-		envURL, accessTok, platformTok, err := getDtEnvironment()
+		envURL, _, platformTok, err := getDtEnvironment()
 		if err != nil {
 			return err
 		}
-		classicTok, err := validateCredentials(envURL, accessTok, platformTok)
-		if err != nil {
+		if _, err := validateCredentials(envURL, "", platformTok); err != nil {
 			return err
 		}
-		c, err := setupClientFromCreds(envURL, classicTok, platformTok)
-		if err != nil {
-			return err
-		}
-		if err := installer.InstallAWS(c.Platform, envURL, platformTok, installDryRun, StartTime.UTC().Format(installer.IngestTimeFormat)); err != nil {
+		if err := awspkg.InstallAWS(envURL, platformTok, installDryRun, StartTime.UTC().Format(installer.IngestTimeFormat)); err != nil {
 			if errors.Is(err, installer.ErrInstallCancelled) {
 				return nil
 			}
