+++ pkg/installer/aws.go
@@ -3,73 +3,19 @@ package installer
 import (
 	"encoding/json"
 	"fmt"
-	"io"
-	"net/http"
 	"os"
 	"os/exec"
 	"strings"
-	"sync"
-
-	"github.com/dynatrace-oss/dtwiz/pkg/client"
-	"github.com/dynatrace-oss/dtwiz/pkg/display"
-	"github.com/dynatrace-oss/dtwiz/pkg/extensions"
 )
 
-// awsTemplateURL points to the latest published Dynatrace CloudFormation
-// template. Using the "latest" channel avoids 403s when a pinned version is
-// removed from S3 and matches dtctl's behaviour.
-const awsTemplateURL = "https://dynatrace-data-acquisition.s3.amazonaws.com/aws/deployment/cfn/latest/da-aws-activation.yaml"
-
-// awsStackConfig holds all values required to render aws.tmpl and drive the
-// CloudFormation deployment.
-type awsStackConfig struct {
-	// StackName is the CloudFormation stack name (unique per account+region).
-	StackName string
-
-	// DynatraceURL is the full Dynatrace Platform environment URL.
-	DynatraceURL string
-
-	// SettingsToken is a platform token (dt0s16.*) with scopes:
-	// settings:objects:write, extensions:configurations:write/read.
-	SettingsToken string
-
-	// IngestToken is a platform token (dt0s16.*) with scopes:
-	// data-acquisition:logs:ingest, data-acquisition:events:ingest.
-	IngestToken string
-
-	// MonitoringConfigID is the UUID of the Dynatrace monitoring configuration.
-	MonitoringConfigID string
-
-	// LogsEnabled controls whether the log-forwarder resources are deployed.
-	LogsEnabled string // "TRUE" | "FALSE"
-
-	// LogsRegions is a comma-separated list of AWS regions for log ingestion.
-	LogsRegions string
-
-	// EventsEnabled controls whether the event-forwarder resources are deployed.
-	EventsEnabled string // "TRUE" | "FALSE"
-
-	// EventsRegions is a comma-separated list of AWS regions for event ingestion.
-	EventsRegions string
-
-	// EventBridgeBusName is the EventBridge bus to consume events from.
-	EventBridgeBusName string
-
-	// EventSources is a comma-separated list of event sources to forward.
-	EventSources string
-
-	// UseCMK controls whether a Customer Managed Key is created for encryption.
-	UseCMK string // "TRUE" | "FALSE"
-}
-
-// isAWSCLIInstalled returns true when the `aws` binary is on PATH.
-func isAWSCLIInstalled() bool {
+// IsAWSCLIInstalled returns true when the `aws` binary is on PATH.
+func IsAWSCLIInstalled() bool {
 	_, err := exec.LookPath("aws")
 	return err == nil
 }
 
-// getAWSCallerInfo returns the AWS account ID and the configured default region.
-func getAWSCallerInfo() (accountID, region string, err error) {
+// GetAWSCallerInfo returns the AWS account ID and the configured default region.
+func GetAWSCallerInfo() (accountID, region string, err error) {
 	// Account ID from STS
 	var sterr strings.Builder
 	cmd := exec.Command("aws", "sts", "get-caller-identity", "--output", "json")
@@ -105,406 +51,4 @@ func getAWSCallerInfo() (accountID, region string, err error) {
 		return accountID, region, nil
 	}
 	return accountID, "us-east-1", nil // safe default
-}
-
-// defaultFeatureSets is the standard set of AWS feature sets forwarded to Dynatrace.
-var defaultFeatureSets = []string{
-	"ApiGateway_essential", "ApplicationELB_essential", "AutoScaling_essential",
-	"CloudFront_essential", "DynamoDB_essential", "EBS_essential", "EC2_essential",
-	"ECR_essential", "ECS_ContainerInsights_essential", "ECS_essential", "EFS_essential",
-	"ELB_essential", "ElastiCache_essential", "Firehose_essential", "Lambda_essential",
-	"NATGateway_essential", "NetworkELB_essential", "PrivateLinkEndpoints_essential",
-	"PrivateLinkServices_essential", "RDS_essential", "Route53_essential", "S3_essential",
-	"SNS_essential", "SQS_essential",
-}
-
-const (
-	daAWSExtension        = "com.dynatrace.extension.da-aws"
-	daAWSExtensionVersion = "1.0.11"
-)
-
-// awsMonitoringConfigValue is used to decode the value field of a da-aws
-// monitoring config item when searching for a matching AWS account.
-type awsMonitoringConfigValue struct {
-	AWS struct {
-		Credentials []struct {
-			AccountID string `json:"accountId"`
-		} `json:"credentials"`
-	} `json:"aws"`
-}
-
-// findExistingMonitoringConfig returns the objectId of an existing da-aws
-// monitoring configuration for the given AWS account, or "" if none is found.
-func findExistingMonitoringConfig(c *client.PlatformClient, accountID string) (string, error) {
-	items, err := extensions.ListMonitoringConfigs(c, daAWSExtension)
-	if err != nil {
-		return "", fmt.Errorf("listing monitoring configs: %w", err)
-	}
-	for _, item := range items {
-		var v awsMonitoringConfigValue
-		if err := json.Unmarshal(item.Value, &v); err != nil {
-			continue
-		}
-		for _, cred := range v.AWS.Credentials {
-			if cred.AccountID == accountID {
-				return item.ObjectID, nil
-			}
-		}
-	}
-	return "", nil
-}
-
-// createDTMonitoringConfig creates a new da-aws monitoring configuration and
-// returns the objectId (UUID) to use as pMonitoringConfigId in CloudFormation.
-func createDTMonitoringConfig(c *client.PlatformClient, accountID, region, version string) (string, error) {
-	desc := fmt.Sprintf("dtwiz — account %s / %s", accountID, region)
-	payload := map[string]interface{}{
-		"scope": "integration-aws",
-		"value": map[string]interface{}{
-			"enabled":           false,
-			"description":       desc,
-			"version":           version,
-			"featureSets":       defaultFeatureSets,
-			"activationContext": "DATA_ACQUISITION",
-			"aws": map[string]interface{}{
-				"deploymentRegion": region,
-				"credentials": []map[string]interface{}{
-					{
-						"description":  desc,
-						"enabled":      false,
-						"connectionId": "*",
-						"accountId":    accountID,
-					},
-				},
-				"regionFiltering":             []string{region},
-				"tagFiltering":                []interface{}{},
-				"tagEnrichment":               []interface{}{},
-				"smartscapeConfiguration":     map[string]interface{}{"enabled": true},
-				"metricsConfiguration":        map[string]interface{}{"enabled": true, "regions": []string{region}},
-				"cloudWatchLogsConfiguration": map[string]interface{}{"enabled": false, "regions": []string{region}},
-				"namespaces":                  []interface{}{},
-				"configurationMode":           "QUICK_START",
-				"deploymentMode":              "AUTOMATED",
-				"deploymentScope":             "SINGLE_ACCOUNT",
-				"manualDeploymentStatus":      "NA",
-				"automatedDeploymentStatus":   "NA",
-			},
-		},
-	}
-	return extensions.CreateMonitoringConfig(c, daAWSExtension, payload)
-}
-
-// downloadAWSTemplate fetches the CloudFormation template from S3 to a
-// temporary file and returns its path.  The caller is responsible for
-// removing the file when done.
-func downloadAWSTemplate() (string, error) {
-	resp, err := http.Get(awsTemplateURL) //nolint:noctx
-	if err != nil {
-		return "", fmt.Errorf("downloading CloudFormation template: %w", err)
-	}
-	defer resp.Body.Close()
-	if resp.StatusCode >= 300 {
-		return "", fmt.Errorf("downloading CloudFormation template: HTTP %d", resp.StatusCode)
-	}
-
-	tmp, err := os.CreateTemp("", "da-aws-activation-*.yaml")
-	if err != nil {
-		return "", fmt.Errorf("creating temp file for CloudFormation template: %w", err)
-	}
-	defer tmp.Close()
-
-	if _, err := io.Copy(tmp, resp.Body); err != nil {
-		_ = os.Remove(tmp.Name())
-		return "", fmt.Errorf("writing CloudFormation template: %w", err)
-	}
-	return tmp.Name(), nil
-}
-
-// maskTokenArgs returns a copy of args with token values truncated to their
-// first 10 characters followed by "***".
-func maskTokenArgs(args []string) []string {
-	tokenPrefixes := []string{"pDtApiToken=", "pDtIngestToken="}
-	out := make([]string, len(args))
-	copy(out, args)
-	for i, a := range out {
-		for _, p := range tokenPrefixes {
-			if strings.HasPrefix(a, p) {
-				val := a[len(p):]
-				if len(val) > 10 {
-					out[i] = p + val[:10] + "***"
-				}
-				break
-			}
-		}
-	}
-	return out
-}
-
-// formatDeployCmd formats the argument slice for display, keeping --flag value
-// pairs on the same line and placing each --parameter-overrides value on its
-// own indented line.
-func formatDeployCmd(args []string) string {
-	var b strings.Builder
-	indent := "\n        "
-	paramIndent := "\n            "
-	inParams := false
-	for i := 0; i < len(args); i++ {
-		a := args[i]
-		if a == "--parameter-overrides" {
-			b.WriteString(" \\")
-			b.WriteString(indent)
-			b.WriteString(a)
-			inParams = true
-			continue
-		}
-		if inParams {
-			b.WriteString(" \\")
-			b.WriteString(paramIndent)
-			b.WriteString(a)
-			continue
-		}
-		if strings.HasPrefix(a, "--") && i+1 < len(args) && !strings.HasPrefix(args[i+1], "--") {
-			b.WriteString(" \\")
-			b.WriteString(indent)
-			b.WriteString(a)
-			b.WriteString(" ")
-			i++
-			b.WriteString(args[i])
-		} else {
-			if b.Len() > 0 {
-				b.WriteString(" ")
-			}
-			b.WriteString(a)
-		}
-	}
-	return b.String()
-}
-
-// buildDeployArgs returns the argument slice for `aws cloudformation deploy`.
-// templateFile must be a local path (aws deploy accepts --template-file only).
-// Each CloudFormation parameter is passed as a separate ParameterKey=Value
-// word in --parameter-overrides so the AWS CLI can correctly handle values
-// that contain commas (e.g. the regions list).
-func buildDeployArgs(cfg awsStackConfig, templateFile string) []string {
-	return []string{
-		"cloudformation", "deploy",
-		"--stack-name", cfg.StackName,
-		"--template-file", templateFile,
-		"--capabilities", "CAPABILITY_NAMED_IAM",
-		"--parameter-overrides",
-		fmt.Sprintf("pDynatraceUrl=%s", cfg.DynatraceURL),
-		fmt.Sprintf("pDtApiToken=%s", cfg.SettingsToken),
-		fmt.Sprintf("pDtIngestToken=%s", cfg.IngestToken),
-		fmt.Sprintf("pMonitoringConfigId=%s", cfg.MonitoringConfigID),
-		fmt.Sprintf("pDtLogsIngestEnabled=%s", cfg.LogsEnabled),
-		fmt.Sprintf("pDtLogsIngestRegions=%s", cfg.LogsRegions),
-		fmt.Sprintf("pDtEventsIngestEnabled=%s", cfg.EventsEnabled),
-		fmt.Sprintf("pDtEventsIngestRegions=%s", cfg.EventsRegions),
-		fmt.Sprintf("pEventBridgeBusName=%s", cfg.EventBridgeBusName),
-		fmt.Sprintf("pEventSources=%s", cfg.EventSources),
-		fmt.Sprintf("pUseCMK=%s", cfg.UseCMK),
-	}
-}
-
-// InstallAWS deploys the Dynatrace AWS Data Acquisition CloudFormation stack.
-//
-// Parameters:
-//   - c:         Platform API client (Extensions monitoring configs via /platform/classic/environment/v2)
-//   - envURL:    Dynatrace Platform environment URL
-//   - token:     platform token (dt0s16.*) used for settings, ingest, and watch
-//   - dryRun:    when true, show what would be done without executing
-//   - startTime: RFC3339 timestamp used as the from-clause for WatchIngest (empty = skip watch)
-func InstallAWS(c *client.PlatformClient, envURL, token string, dryRun bool, startTime string) error {
-	sep := strings.Repeat("─", 60)
-
-	fmt.Println()
-	display.ColorMessage.Println("  Dynatrace AWS CloudFormation Integration")
-	fmt.Println()
-
-	// ── Validate parameters ──────────────────────────────────────────────────
-
-	stackName := "dynatrace-data-acquisition"
-	dynatraceURL := envURL
-	if dynatraceURL == "" {
-		return fmt.Errorf("Dynatrace environment URL is required (--environment or DT_ENVIRONMENT)") //nolint:staticcheck // ST1005: keep brand capitalization
-	}
-
-	if token == "" {
-		return fmt.Errorf("platform token is required (--platform-token or DT_PLATFORM_TOKEN)")
-	}
-	settingsToken := token
-	ingestToken := token
-
-	// ── Preflight ─────────────────────────────────────────────────────────────
-
-	if !isAWSCLIInstalled() {
-		return fmt.Errorf("AWS CLI not found — install it from https://docs.aws.amazon.com/cli/latest/userguide/install-cliv2.html")
-	}
-
-	// ── Render preview ────────────────────────────────────────────────────────
-
-	fmt.Printf("\n  Fetching AWS account info...\n")
-	accountID, region, err := getAWSCallerInfo()
-	if err != nil {
-		return fmt.Errorf("fetching AWS caller info: %w", err)
-	}
-	fmt.Printf("  AWS account: %s  region: %s\n", accountID, region)
-	fmt.Printf("  Template: %s\n", awsTemplateURL)
-
-	// Monitoring config ID is resolved after confirmation; show placeholder in preview.
-	cfg := awsStackConfig{
-		StackName:          stackName,
-		DynatraceURL:       strings.TrimRight(AppsURL(dynatraceURL), "/"),
-		SettingsToken:      settingsToken,
-		IngestToken:        ingestToken,
-		MonitoringConfigID: "(auto-assigned)",
-		LogsEnabled:        "TRUE",
-		LogsRegions:        region,
-		EventsEnabled:      "TRUE",
-		EventsRegions:      region,
-		EventBridgeBusName: "default",
-		EventSources:       "aws.health",
-		UseCMK:             "FALSE",
-	}
-	deployArgs := buildDeployArgs(cfg, "/tmp/da-aws-activation.yaml")
-
-	fmt.Println()
-	fmt.Printf("  %s\n", sep)
-	display.ColorMessage.Println("  Command to be executed:")
-	fmt.Printf("  %s\n", sep)
-	fmt.Printf("    aws %s\n", formatDeployCmd(maskTokenArgs(deployArgs)))
-	fmt.Printf("  %s\n\n", sep)
-
-	if dryRun {
-		fmt.Println("  [dry-run] No changes were made.")
-		return nil
-	}
-
-	// ── Confirm ───────────────────────────────────────────────────────────────
-
-	ok, err := confirmProceed("  Proceed with installation?")
-	if err != nil {
-		return fmt.Errorf("reading confirmation: %w", err)
-	}
-	if !ok {
-		fmt.Println("  Deployment cancelled.")
-		return ErrInstallCancelled
-	}
-	fmt.Println()
-
-	// ── Dynatrace setup (extension + monitoring config) ───────────────────────
-
-	if err := extensions.InstallExtension(c, daAWSExtension, daAWSExtensionVersion, true); err != nil {
-		return fmt.Errorf("installing extension %s: %w", daAWSExtension, err)
-	}
-
-	monitoringConfigID, err := findExistingMonitoringConfig(c, accountID)
-	if err != nil {
-		return fmt.Errorf("looking up monitoring configuration: %w", err)
-	}
-	if monitoringConfigID != "" {
-		fmt.Printf("  Monitoring config: found existing %s\n", monitoringConfigID)
-	} else {
-		extVersion, vErr := extensions.GetLatestInstalledVersion(c, daAWSExtension)
-		if vErr != nil {
-			fmt.Printf("  Warning: could not resolve installed extension version (%s); falling back to %s\n", vErr, daAWSExtensionVersion)
-			extVersion = daAWSExtensionVersion
-		}
-		fmt.Printf("  Creating Dynatrace monitoring configuration (extension %s)...\n", extVersion)
-		monitoringConfigID, err = createDTMonitoringConfig(c, accountID, region, extVersion)
-		if err != nil {
-			return fmt.Errorf("creating monitoring configuration: %w", err)
-		}
-		fmt.Printf("  Monitoring config: created %s\n", monitoringConfigID)
-	}
-
-	cfg.MonitoringConfigID = monitoringConfigID
-
-	// ── Deploy ────────────────────────────────────────────────────────────────
-
-	fmt.Printf("  Downloading CloudFormation template...\n")
-	tmplFile, err := downloadAWSTemplate()
-	if err != nil {
-		return err
-	}
-	defer os.Remove(tmplFile)
-
-	realArgs := buildDeployArgs(cfg, tmplFile)
-
-	// statusCh carries deployment progress messages into the watch display.
-	statusCh := make(chan string, 4)
-
-	// Start CFN deploy in the background immediately — it takes several minutes
-	// and produces no meaningful intermediate output.
-	var wg sync.WaitGroup
-	var deployErr error
-
-	wg.Add(1)
-	go func() {
-		defer wg.Done()
-		statusCh <- fmt.Sprintf("CloudFormation stack %q deploying... (this may take a few minutes)", cfg.StackName)
-		if err := RunCommandQuiet("aws", realArgs...); err != nil {
-			deployErr = fmt.Errorf("CloudFormation deployment failed: %w", err)
-			statusCh <- fmt.Sprintf("CloudFormation deployment failed: %s", err)
-			return
-		}
-		statusCh <- fmt.Sprintf("CloudFormation stack %q deployed successfully.", cfg.StackName)
-
-		statusCh <- "Enabling Dynatrace AWS monitoring configuration..."
-		if err := enableAWSMonitoringConfig(c, monitoringConfigID); err != nil {
-			deployErr = fmt.Errorf("enabling monitoring configuration: %w", err)
-			statusCh <- fmt.Sprintf("Enabling monitoring configuration failed: %s", err)
-			return
-		}
-		statusCh <- "Dynatrace AWS monitoring configuration enabled."
-	}()
-
-	// Run Lambda instrumentation on the main thread — it is quick but produces
-	// a lot of output, so let it finish before handing the terminal to watch.
-	lambdaErr := InstallAWSLambda(envURL, token, false, false)
-	if lambdaErr != nil {
-		fmt.Printf("\n  Warning: Lambda instrumentation encountered an error: %s\n", lambdaErr)
-		fmt.Println("  You can retry with: dtwiz install aws-lambda")
-	}
-
-	// Start watch after Lambda output is done — CFN deploy is still running
-	// in the background and will send its result into statusCh. Scope the
-	// cloud-platform signal queries to this AWS account to filter out noise
-	// from other accounts in the same tenant.
-	if startTime != "" && token != "" {
-		WatchIngestAWS(envURL, token, startTime, statusCh, accountID)
-	}
-
-	wg.Wait()
-	close(statusCh)
-
-	if deployErr != nil {
-		return deployErr
-	}
-	return nil
-}
-
-// enableAWSMonitoringConfig flips the da-aws monitoring configuration and all
-// its credentials to enabled=true. Mirrors `dtctl enable aws monitoring`:
-// without this step the CloudFormation stack is deployed but Dynatrace will
-// not actually collect anything.
-//
-// The AWS-specific mutation of the value payload lives here; the underlying
-// GET/PUT against the Extensions v2 API is provided by pkg/extensions.
-func enableAWSMonitoringConfig(c *client.PlatformClient, id string) error {
-	cfg, err := extensions.GetMonitoringConfig(c, daAWSExtension, id)
-	if err != nil {
-		return err
-	}
-	cfg.Value["enabled"] = true
-	if aws, ok := cfg.Value["aws"].(map[string]interface{}); ok {
-		if creds, ok := aws["credentials"].([]interface{}); ok {
-			for _, cred := range creds {
-				if m, ok := cred.(map[string]interface{}); ok {
-					m["enabled"] = true
-				}
-			}
-		}
-	}
-	return extensions.UpdateMonitoringConfig(c, daAWSExtension, id, cfg)
 }
