+++ pkg/installer/aws/uninstall.go
@@ -1,22 +1,30 @@
-package installer
+package aws
 
 import (
 	"fmt"
 
-	"github.com/dynatrace-oss/dtwiz/pkg/client"
 	"github.com/dynatrace-oss/dtwiz/pkg/display"
-	"github.com/dynatrace-oss/dtwiz/pkg/extensions"
+	"github.com/dynatrace-oss/dtwiz/pkg/installer"
 )
 
 // UninstallAWS removes the Dynatrace AWS CloudFormation stack and the
 // associated Dynatrace monitoring configuration.
 //
 // Parameters:
-//   - c:      Platform API client (used for Extensions API calls)
-//   - envURL: Dynatrace environment URL (reserved for future use)
-//   - dryRun: when true, show what would be done without executing
-func UninstallAWS(c *client.PlatformClient, envURL string, dryRun bool) error {
-	if !isAWSCLIInstalled() {
+//   - envURL:  Dynatrace Platform environment URL
+//   - token:   platform token (dt0s16.*) used for Extensions API calls
+//   - dryRun:  when true, show what would be done without executing
+func UninstallAWS(envURL, token string, dryRun bool) error {
+	dtc, err := newSDKDTClient(envURL, token)
+	if err != nil {
+		return err
+	}
+	return uninstallAWSWithClient(dryRun, dtc)
+}
+
+// uninstallAWSWithClient is the testable core; dtclient is injected.
+func uninstallAWSWithClient(dryRun bool, dtc dtclient) error {
+	if !installer.IsAWSCLIInstalled() {
 		return fmt.Errorf("AWS CLI not found — install it from https://docs.aws.amazon.com/cli/latest/userguide/install-cliv2.html")
 	}
 
@@ -25,7 +33,7 @@ func UninstallAWS(c *client.PlatformClient, envURL string, dryRun bool) error {
 	fmt.Println()
 
 	fmt.Printf("  Fetching AWS account info...\n")
-	accountID, region, err := getAWSCallerInfo()
+	accountID, region, err := installer.GetAWSCallerInfo()
 	if err != nil {
 		return fmt.Errorf("fetching AWS caller info: %w", err)
 	}
@@ -34,7 +42,7 @@ func UninstallAWS(c *client.PlatformClient, envURL string, dryRun bool) error {
 	stackName := "dynatrace-data-acquisition"
 
 	fmt.Printf("  Looking up Dynatrace AWS monitoring configuration...\n")
-	monitoringConfigID, err := findExistingMonitoringConfig(c, accountID)
+	monitoringConfigID, err := dtc.findExistingMonitoringConfig(accountID)
 	if err != nil {
 		return fmt.Errorf("looking up monitoring configuration: %w", err)
 	}
@@ -52,38 +60,26 @@ func UninstallAWS(c *client.PlatformClient, envURL string, dryRun bool) error {
 	}
 	fmt.Println()
 
-	if dryRun {
-		fmt.Println("  [dry-run] No changes were made.")
-		return nil
-	}
-
-	ok, err := confirmProceed("  Proceed with uninstall?")
-	if err != nil {
-		return fmt.Errorf("reading confirmation: %w", err)
-	}
-	if !ok {
-		fmt.Println("  Uninstall cancelled.")
-		return nil
+	if proceed, err := installer.ShouldProceed(dryRun, "Uninstall"); !proceed {
+		return err
 	}
 	fmt.Println()
 
-	// Step 1: Delete the CloudFormation stack.
 	fmt.Printf("  Step 1: Deleting CloudFormation stack %q...\n", stackName)
-	if err := RunCommand("aws", "cloudformation", "delete-stack",
+	if err := installer.RunCommand("aws", "cloudformation", "delete-stack",
 		"--stack-name", stackName, "--region", region); err != nil {
 		return fmt.Errorf("deleting CloudFormation stack: %w", err)
 	}
 	fmt.Printf("  Waiting for stack deletion to complete (this may take several minutes)...\n")
-	if err := RunCommand("aws", "cloudformation", "wait", "stack-delete-complete",
+	if err := installer.RunCommand("aws", "cloudformation", "wait", "stack-delete-complete",
 		"--stack-name", stackName, "--region", region); err != nil {
 		return fmt.Errorf("waiting for stack deletion: %w", err)
 	}
 	fmt.Printf("  Stack %q deleted.\n", stackName)
 
-	// Step 2: Delete the Dynatrace monitoring configuration.
 	if monitoringConfigID != "" {
 		fmt.Printf("\n  Step 2: Deleting Dynatrace AWS monitoring configuration %s...\n", monitoringConfigID)
-		if err := extensions.DeleteMonitoringConfig(c, daAWSExtension, monitoringConfigID); err != nil {
+		if err := dtc.deleteMonitoringConfig(monitoringConfigID); err != nil {
 			return fmt.Errorf("deleting monitoring config: %w", err)
 		}
 		fmt.Printf("  Monitoring configuration %s deleted.\n", monitoringConfigID)
